Fix: windows string paths.

+++ Appcoins Unity/Assets/AppcoinsUnity/Scripts/BashUtils.cs
@@ -230,27 +230,35 @@ namespace Aptoide.AppcoinsUnity
 
         private void CreateBatchFileToExecuteCommand(int buildPhase, string cmd, string cmdArgs, bool debugMode, string path)
         {
-            // Delete all temporary files.
-            if (File.Exists(Application.dataPath + "\\AppcoinsUnity\\Tools\\ProcessCompleted.out"))
+            
+            if(!Directory.Exists(Application.dataPath + "\\AppcoinsUnity\\Tools"))
             {
-                File.Delete(Application.dataPath + "\\AppcoinsUnity\\Tools\\ProcessCompleted.out");
+                Directory.CreateDirectory(Application.dataPath + "\\AppcoinsUnity\\Tools");
             }
-
-            if (File.Exists(Application.dataPath + "\\AppcoinsUnity\\Tools\\ProcessError.out"))
+            
+            else
             {
-                File.Delete(Application.dataPath + "\\AppcoinsUnity\\Tools\\ProcessError.out");
-            }
+                // Delete all temporary files.
+                if (File.Exists(Application.dataPath + "\\AppcoinsUnity\\Tools\\ProcessCompleted.out"))
+                {
+                    File.Delete(Application.dataPath + "\\AppcoinsUnity\\Tools\\ProcessCompleted.out");
+                }
 
-            if (File.Exists(Application.dataPath + "\\AppcoinsUnity\\Tools\\BashCommand.bat"))
-            {
-                File.Delete(Application.dataPath + "\\AppcoinsUnity\\Tools\\BashCommand.bat");
-            }
+                if (File.Exists(Application.dataPath + "\\AppcoinsUnity\\Tools\\ProcessError.out"))
+                {
+                    File.Delete(Application.dataPath + "\\AppcoinsUnity\\Tools\\stderr.out");
+                }
 
-            if (File.Exists(Application.dataPath + "\\AppcoinsUnity\\Tools\\BashCommand.sh"))
-            {
-                File.Delete(Application.dataPath + "\\AppcoinsUnity\\Tools\\BashCommand.sh");
-            }
+                if (File.Exists(Application.dataPath + "\\AppcoinsUnity\\Tools\\BashCommand.bat"))
+                {
+                    File.Delete(Application.dataPath + "\\AppcoinsUnity\\Tools\\BashCommand.bat");
+                }
 
+                if (File.Exists(Application.dataPath + "\\AppcoinsUnity\\Tools\\BashCommand.sh"))
+                {
+                    File.Delete(Application.dataPath + "\\AppcoinsUnity\\Tools\\BashCommand.sh");
+                }
+            }
 
             StreamWriter writer = new StreamWriter(Application.dataPath + "\\AppcoinsUnity\\Tools\\BashCommand.bat", false);
